refactor: 유효성 검사 메서드 리팩토링
